requirements.txt
black
pylint
pytest
click
python-telegram-bot
python-decouple
pandas
numpy
mypy
pydantic
